hw1/requirements.txt
matplotlib==3.8.2
numpy==1.26.2
opencv_python==4.8.1.78
Pillow==10.1.0